Took folders out of Developer folder, causing error, add GET and POST functions
