+++ db/db.json
@@ -1,0 +1,18 @@
+[
+    {
+        "title": "Test Title",
+        "text": "Test text"
+    },
+    {
+        "title": "asdfasd",
+        "text": "sdafasdfasd"
+    },
+    {
+        "title": "sadfsd",
+        "text": "sdfasdf"
+    },
+    {
+        "title": "Heres a note",
+        "text": "theres a note"
+    }
+]
