+++ server.js
@@ -1,8 +1,9 @@
 const express = require('express');
 const fs = require('fs');
 const path = require('path');
-const { readFromFile, readAndAppend } = require('./Develop/helpers/fsUtils');
-const { notes } = require('./Develop/db/db.json');
+const util = require('util');
+const { notes } = require('./db/db.json');
+
 
 const PORT = process.env.port || 3001;
 
@@ -15,59 +16,73 @@ app.use(express.urlencoded({ extended: true }));
 
 app.use(express.static('public'));
 
-// function createNote (body, notesAdd) {
-//     const note = body; 
-//     notesAdd.push(note); 
-
-//     // path to write file 
-//     fs.writeFileSync(
-//         path.join(__dirname, './Develop/db/db.json'),
-//         JSON.stringify({ notes : notesAdd }, null, 2)
-//     );
-//     return note; 
-// };
-
-app.get('/notes', (req, res) => {
-    // res.json(notes); 
-    readFromFile('./Develop/db/db.json').then((data) => res.json(JSON.parse(data)));
-});
-
-app.post('/notes', (req, res) => {
-    console.log(req.body);
-
-    const  { title, text } = req.body;
-
-    if (req.body) {
-        const newNote = {
-            title,
-            text,
-        };
-
-        readAndAppend(newNote, './Develop/db/db.json');
-        res.json('Note successfully added');
-    } else {
-        res.errored('Error in adding note');
-    }
-})
-
-// app.post('/notes', (req, res) => {
-//     req.body.id = notes.length.toString(); 
-//     const note = createNote(req.body, notes); 
-
-//     res.json(note);
-//     }
-// );
-
-
 // GET Route for homepage
 app.get('/', (req, res) =>
-    res.sendFile(path.join(__dirname, '/Develop/public/index.html'))
+    res.sendFile(path.join(__dirname, '/public/index.html'))
 );
 
 // GET Route for notes page
 app.get('/notes', (req, res) =>
-    res.sendFile(path.join(__dirname, '/Develop/public/notes.html'))
+    res.sendFile(path.join(__dirname, '/public/notes.html'))
 );
+
+const readFromFile = util.promisify(fs.readFile);
+
+/**
+ *  Function to write data to the JSON file given a destination and some content
+ *  @param {string} destination The file you want to write to.
+ *  @param {object} content The content you want to write to the file.
+ *  @returns {void} Nothing
+ */
+const writeToFile = (destination, content) =>
+fs.writeFile(destination, JSON.stringify(content, null, 4), (err) =>
+    err ? console.error(err) : console.info(`\nData written to ${destination}`)
+);
+
+/**
+ *  Function to read data from a given a file and append some content
+ *  @param {object} content The content you want to append to the file.
+ *  @param {string} file The path to the file you want to save to.
+ *  @returns {void} Nothing
+ */
+const readAndAppend = (content, file) => {
+    console.log(content, file);
+fs.readFile(file, 'utf8', (err, data) => {
+    if (err) {
+    console.error(err);
+    } else {
+    const parsedData = JSON.parse(data);
+    parsedData.push(content);
+    writeToFile(file, parsedData);
+    }
+});
+};
+
+// GET Route for retrieving all the tips
+app.get('/api/notes', (req, res) => {
+    console.info(`${req.method} request received for notes`);
+    readFromFile('./db/db.json').then((data) => res.json(JSON.parse(data)));
+});
+
+  // POST Route for a new UX/UI tip
+app.post('/api/notes', (req, res) => {
+    console.info(`${req.method} request received to add a note`);
+
+    const { title, text } = req.body;
+
+    if (req.body) {
+    const newNote = {
+        title,
+        text,
+    };
+
+    readAndAppend(newNote, './db/db.json');
+    res.json(`Note added successfully 🚀`);
+    } else {
+    res.error('Error in adding note');
+    }
+});
+
 
 app.listen(PORT, () =>
     console.log(`App listening at http://localhost:${PORT} 🚀`)
